@@ -1,5 +1,5 @@
 [package]
-name = "tokio-sando"
+name = "sando"
 description = "A simple proxy server implementation"
 version = "0.1.0"
 authors = ["Chun-Min Chang <chun.m.chang@gmail.com>"]
@@ -7,15 +7,15 @@ license = "MIT"
 repository = "https://github.com/ChunMinChang/tokio-sando"
 keywords = ["http", "https", "tunnel", "proxy", "connect"]
 categories = ["web-programming::http-server", "network-programming"]
-edition = "2018"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-argh = "0.1"
-regex = "1"
-tokio = { version = "1.0", features = [ "full" ] }
-tokio-native-tls = "0.3"
+argh = "0.1.9"
+regex = "1.6.0"
+tokio = { version = "1.21.2", features = [ "full" ] }
+tokio-native-tls = "0.3.0"
 
 [dev-dependencies]
-tokio-test = "0.4"
+tokio-test = "0.4.2"
